Revert "video model rename"
This reverts commit ab49be28170d391b66a076444e5f38c2b8efd902.
